requirements.txt
djangorestframework
python-dotenv
psycopg2-binary
python-slugify